+++ src/Select/Select.tsx
@@ -43,18 +43,24 @@ export interface SelectInputComponentProps {
   onClick: ComponentProps<typeof SelectInput>['onClick']
 }
 
+type CloseDropdown = () => void
+
+type ChildrenCallback = {
+  close: CloseDropdown
+}
+
 export interface SelectProps<V> extends Pick<DropdownProps, 'placement'> {
   selected: V
   inputValueAccessor?: (item: V) => string
-  InputComponent?: React.ComponentType<SelectInputComponentProps>
+  inputComponent?: (props: SelectInputComponentProps) => JSX.Element
   data?: V extends SelectListData ? V[] : never
   onChange?: V extends SelectListData ? ((item: V) => void) : never
   name?: string
   placeholder?: string
   disabled?: boolean
-  onOutsideClick?: (isOpen: boolean, close: () => void) => void
+  onOutsideClick?: (isOpen: boolean, close: CloseDropdown) => void
   useInputComponentWidth?: boolean
-  children?: React.ReactElement
+  children?: ((props: ChildrenCallback) => React.ReactNode) | React.ReactElement
   open?: boolean | null
   defaultOpen?: boolean
   onToggle?: (open: boolean) => void
@@ -72,7 +78,7 @@ export function Select <V extends SelectData> ({
   onChange,
   useInputComponentWidth = true,
   onOutsideClick,
-  InputComponent = SelectInput,
+  inputComponent,
   children,
   py = 2,
   disabled = false,
@@ -97,7 +103,11 @@ export function Select <V extends SelectData> ({
       open={isOpen}
       placement={placement}
       useTriggerWidth={useInputComponentWidth}
-      trigger={<InputComponent {...inputProps} />}
+      trigger={
+        inputComponent
+          ? inputComponent(inputProps)
+          : <SelectInput {...inputProps} />
+      }
       onOutsideClick={() => {
         if (onOutsideClick) {
           onOutsideClick(isOpen, close)
@@ -106,15 +116,21 @@ export function Select <V extends SelectData> ({
         }
       }}>
       <OptionsWrapper fullWidth={useInputComponentWidth}>
-        {children || (
-          <SelectList<V extends SelectListData ? V : never>
-            value={selected as V extends SelectListData ? V : never}
-            data={data as any[] || []}
-            onChange={value => {
-              onChange && onChange(value)
-              close()
-            }} />
-        )}
+        {
+          typeof children === 'function'
+            ? children({ close })
+            : (
+              children || (
+                <SelectList<V extends SelectListData ? V : never>
+                  value={selected as V extends SelectListData ? V : never}
+                  data={data as any[] || []}
+                  onChange={(value) => {
+                    onChange && onChange(value)
+                    close()
+                  }} />
+              )
+            )
+        }
       </OptionsWrapper>
     </Dropdown>
   )